# Replace publish.Reporter with model.BatchProcessor