--- a/beater/beater.go
+++ b/beater/beater.go
@@ -46,6 +46,7 @@ import (
 	"github.com/elastic/apm-server/elasticsearch"
 	"github.com/elastic/apm-server/ingest/pipeline"
 	logs "github.com/elastic/apm-server/log"
+	"github.com/elastic/apm-server/model"
 	"github.com/elastic/apm-server/model/modelprocessor"
 	"github.com/elastic/apm-server/publish"
 	"github.com/elastic/apm-server/sampling"
@@ -344,17 +345,6 @@ func (s *serverRunner) run() error {
 	// Send config to telemetry.
 	recordAPMServerConfig(s.config)
 
-	runServer := runServer
-	if s.tracerServer != nil {
-		runServer = runServerWithTracerServer(runServer, s.tracerServer, s.tracer)
-	}
-	if s.wrapRunServer != nil {
-		// Wrap runServer function, enabling injection of
-		// behaviour into the processing/reporting pipeline.
-		runServer = s.wrapRunServer(runServer)
-	}
-	runServer = s.wrapRunServerWithPreprocessors(runServer)
-
 	transformConfig, err := newTransformConfig(s.beat.Info, s.config)
 	if err != nil {
 		return err
@@ -379,22 +369,38 @@ func (s *serverRunner) run() error {
 	}
 	defer publisher.Stop(s.backgroundContext)
 
+	// Create the runServer function. We start with newBaseRunServer, and then
+	// wrap depending on the configuration in order to inject behaviour.
 	reporter := publisher.Send
+	runServer := newBaseRunServer(reporter)
+	if s.tracerServer != nil {
+		runServer = runServerWithTracerServer(runServer, s.tracerServer, s.tracer)
+	}
+	if s.wrapRunServer != nil {
+		// Wrap runServer function, enabling injection of
+		// behaviour into the processing/reporting pipeline.
+		runServer = s.wrapRunServer(runServer)
+	}
+	runServer = s.wrapRunServerWithPreprocessors(runServer)
+
+	var batchProcessor model.BatchProcessor = &reporterBatchProcessor{reporter}
 	if !s.config.Sampling.KeepUnsampled {
 		// The server has been configured to discard unsampled
 		// transactions. Make sure this is done just before calling
 		// the publisher to avoid affecting aggregations.
-		reporter = sampling.NewDiscardUnsampledReporter(reporter)
+		batchProcessor = wrapBatchProcessor(
+			batchProcessor, sampling.NewDiscardUnsampledBatchProcessor(),
+		)
 	}
 
 	if err := runServer(s.runServerContext, ServerParams{
-		Info:      s.beat.Info,
-		Config:    s.config,
-		Managed:   s.beat.Manager != nil && s.beat.Manager.Enabled(),
-		Namespace: s.namespace,
-		Logger:    s.logger,
-		Tracer:    s.tracer,
-		Reporter:  reporter,
+		Info:           s.beat.Info,
+		Config:         s.config,
+		Managed:        s.beat.Manager != nil && s.beat.Manager.Enabled(),
+		Namespace:      s.namespace,
+		Logger:         s.logger,
+		Tracer:         s.tracer,
+		BatchProcessor: batchProcessor,
 	}); err != nil {
 		return err
 	}
@@ -402,7 +408,7 @@ func (s *serverRunner) run() error {
 }
 
 func (s *serverRunner) wrapRunServerWithPreprocessors(runServer RunServerFunc) RunServerFunc {
-	var processors []transform.Processor
+	var processors []model.BatchProcessor
 	if s.config.DefaultServiceEnvironment != "" {
 		processors = append(processors, &modelprocessor.SetDefaultServiceEnvironment{
 			DefaultServiceEnvironment: s.config.DefaultServiceEnvironment,
@@ -567,7 +573,7 @@ func runServerWithTracerServer(runServer RunServerFunc, tracerServer *tracerServ
 	return func(ctx context.Context, args ServerParams) error {
 		g, ctx := errgroup.WithContext(ctx)
 		g.Go(func() error {
-			return tracerServer.serve(ctx, args.Reporter)
+			return tracerServer.serve(ctx, args.BatchProcessor)
 		})
 		g.Go(func() error {
 			return runServer(ctx, args)
@@ -606,24 +612,36 @@ func newSourcemapStore(beatInfo beat.Info, cfg *config.SourceMapping) (*sourcema
 }
 
 // WrapRunServerWithProcessors wraps runServer such that it wraps args.Reporter
-// with a function that transformables are first passed through the given
-// processors in order.
-func WrapRunServerWithProcessors(runServer RunServerFunc, processors ...transform.Processor) RunServerFunc {
+// with a function that event batches are first passed through the given processors
+// in order.
+func WrapRunServerWithProcessors(runServer RunServerFunc, processors ...model.BatchProcessor) RunServerFunc {
 	if len(processors) == 0 {
 		return runServer
 	}
 	return func(ctx context.Context, args ServerParams) error {
-		origReporter := args.Reporter
-		args.Reporter = func(ctx context.Context, req publish.PendingReq) error {
-			var err error
-			for _, p := range processors {
-				req.Transformables, err = p.ProcessTransformables(ctx, req.Transformables)
-				if err != nil {
-					return err
-				}
-			}
-			return origReporter(ctx, req)
-		}
+		args.BatchProcessor = wrapBatchProcessor(args.BatchProcessor, processors...)
 		return runServer(ctx, args)
 	}
+}
+
+func wrapBatchProcessor(final model.BatchProcessor, pre ...model.BatchProcessor) model.BatchProcessor {
+	return model.ProcessBatchFunc(func(ctx context.Context, batch *model.Batch) error {
+		for _, p := range pre {
+			if err := p.ProcessBatch(ctx, batch); err != nil {
+				return err
+			}
+		}
+		return final.ProcessBatch(ctx, batch)
+	})
+}
+
+type disablePublisherTracingKey struct{}
+
+type reporterBatchProcessor struct {
+	reporter publish.Reporter
+}
+
+func (p *reporterBatchProcessor) ProcessBatch(ctx context.Context, batch *model.Batch) error {
+	disableTracing, _ := ctx.Value(disablePublisherTracingKey{}).(bool)
+	return p.reporter(ctx, publish.PendingReq{Transformable: batch, Trace: !disableTracing})
 }