--- a/processor/otel/consumer.go
+++ b/processor/otel/consumer.go
@@ -51,7 +51,6 @@ import (
 
 	logs "github.com/elastic/apm-server/log"
 	"github.com/elastic/apm-server/model"
-	"github.com/elastic/apm-server/publish"
 )
 
 const (
@@ -68,7 +67,7 @@ const (
 type Consumer struct {
 	stats consumerStats
 
-	Reporter publish.Reporter
+	Processor model.BatchProcessor
 }
 
 // ConsumerStats holds a snapshot of statistics about data consumption.
@@ -95,10 +94,7 @@ func (c *Consumer) Stats() ConsumerStats {
 // converting into Elastic APM events and reporting to the Elastic APM schema.
 func (c *Consumer) ConsumeTraces(ctx context.Context, traces pdata.Traces) error {
 	batch := c.convert(traces)
-	return c.Reporter(ctx, publish.PendingReq{
-		Transformables: batch.Transformables(),
-		Trace:          true,
-	})
+	return c.Processor.ProcessBatch(ctx, batch)
 }
 
 func (c *Consumer) convert(td pdata.Traces) *model.Batch {