## Motivation/summary

Instead of all code being passed a `publish.Reporter` function, which deals with "Transformables", we now introduce make available a `model.BatchProcessor`, which deals with the concrete type `model.Batch`. Since `BatchProcessor` deals with `model.Batches`, it enables greater reliance on static typing and reduces the need for type assertions.

The beater provides a `BatchProcessor` that invokes `publish.Reporter`; this `BatchProcessor` can be wrapped for processing, aggregation, sampling, etc. To simplify the publishing adapter, `model.Batch` now implements the `Transformable` interface directly. All existing processors (aggregations, tail-based sampling, etc.) have been updated to implement the `BatchProcessor` interface.

The `publish.Reporter` is still used for publishing onboarding docs and sourcemaps. Since these are internal to the beater package (and child packages), we no longer exposed the reporter to any other packages.

There is a second, only tangentially related, commit in this PR which optimises `processor/stream.Processor.IdentifyEventType`. This method was causing more allocations than it should (i.e. more than zero). I included the commit to pay for some of the (temporary) loss of performance incurred by the refactoring.

## How to test these changes

Non functional change, run the tests.

## Related issues

Closes https://github.com/elastic/apm-server/issues/4877