@axw 2021-02-27
Modifying processor/stream benchmarks to transform events, comparing to the master branch:

```
name                                                               old time/op    new time/op    delta
BackendProcessor/errors.ndjson/NoRateLimit-12                         133µs ± 1%     136µs ± 2%    ~     (p=0.200 n=3+3)
BackendProcessor/errors.ndjson/WithRateLimit-12                       135µs ± 1%     138µs ± 3%    ~     (p=1.000 n=3+3)
BackendProcessor/errors_2.ndjson/NoRateLimit-12                       103µs ± 1%     104µs ± 1%    ~     (p=0.700 n=3+3)
BackendProcessor/errors_2.ndjson/WithRateLimit-12                     105µs ± 0%     105µs ± 3%    ~     (p=0.700 n=3+3)
BackendProcessor/errors_rum.ndjson/NoRateLimit-12                    38.4µs ± 8%    32.6µs ± 4%    ~     (p=0.100 n=3+3)
BackendProcessor/errors_rum.ndjson/WithRateLimit-12                  41.9µs ±15%    35.0µs ± 3%    ~     (p=0.100 n=3+3)
BackendProcessor/events.ndjson/NoRateLimit-12                         186µs ± 6%     182µs ± 0%    ~     (p=1.000 n=3+3)
BackendProcessor/events.ndjson/WithRateLimit-12                       180µs ± 1%     192µs ± 4%    ~     (p=0.100 n=3+3)
BackendProcessor/experimental.ndjson/NoRateLimit-12                  36.6µs ± 7%    31.9µs ± 6%    ~     (p=0.100 n=3+3)
BackendProcessor/experimental.ndjson/WithRateLimit-12                39.2µs ± 9%    32.3µs ± 2%    ~     (p=0.100 n=3+3)
BackendProcessor/heavy.ndjson/NoRateLimit-12                         30.8ms ±11%    18.0ms ± 2%    ~     (p=0.100 n=3+3)
BackendProcessor/heavy.ndjson/WithRateLimit-12                       29.0ms ±10%    20.2ms ±10%    ~     (p=0.100 n=3+3)
BackendProcessor/invalid-event-type.ndjson/NoRateLimit-12            6.78µs ± 1%    7.59µs ± 2%    ~     (p=0.100 n=3+3)
BackendProcessor/invalid-event-type.ndjson/WithRateLimit-12          8.13µs ± 1%    8.93µs ± 1%    ~     (p=0.100 n=3+3)
BackendProcessor/invalid-event.ndjson/NoRateLimit-12                 44.4µs ± 3%    44.7µs ± 5%    ~     (p=1.000 n=3+3)
BackendProcessor/invalid-event.ndjson/WithRateLimit-12               45.0µs ± 4%    46.2µs ± 1%    ~     (p=0.400 n=3+3)
BackendProcessor/invalid-json-event.ndjson/NoRateLimit-12            17.9µs ± 8%    18.3µs ± 3%    ~     (p=1.000 n=3+3)
BackendProcessor/invalid-json-event.ndjson/WithRateLimit-12          19.2µs ± 3%    20.2µs ± 3%    ~     (p=0.100 n=3+3)
BackendProcessor/invalid-json-metadata.ndjson/NoRateLimit-12         19.1µs ± 2%    22.9µs ±13%    ~     (p=0.100 n=3+3)
BackendProcessor/invalid-json-metadata.ndjson/WithRateLimit-12       19.2µs ± 0%    24.8µs ±11%    ~     (p=0.100 n=3+3)
BackendProcessor/invalid-metadata-2.ndjson/NoRateLimit-12            3.80µs ± 2%    4.33µs ± 5%    ~     (p=0.100 n=3+3)
BackendProcessor/invalid-metadata-2.ndjson/WithRateLimit-12          3.81µs ± 1%    4.07µs ± 1%    ~     (p=0.100 n=3+3)
BackendProcessor/invalid-metadata.ndjson/NoRateLimit-12              3.96µs ± 1%    4.28µs ± 7%    ~     (p=0.100 n=3+3)
BackendProcessor/invalid-metadata.ndjson/WithRateLimit-12            3.91µs ± 0%    4.64µs ± 6%    ~     (p=0.100 n=3+3)
BackendProcessor/metadata-null-values.ndjson/NoRateLimit-12          12.5µs ± 1%    13.3µs ± 1%    ~     (p=0.100 n=3+3)
BackendProcessor/metadata-null-values.ndjson/WithRateLimit-12        14.5µs ± 2%    14.7µs ± 1%    ~     (p=0.400 n=3+3)
BackendProcessor/metadata.ndjson/NoRateLimit-12                      11.2µs ± 1%    12.3µs ± 1%    ~     (p=0.100 n=3+3)
BackendProcessor/metadata.ndjson/WithRateLimit-12                    12.6µs ± 0%    13.9µs ± 1%    ~     (p=0.100 n=3+3)
BackendProcessor/metricsets.ndjson/NoRateLimit-12                    58.4µs ± 1%    59.5µs ± 6%    ~     (p=0.700 n=3+3)
BackendProcessor/metricsets.ndjson/WithRateLimit-12                  60.1µs ± 4%    60.7µs ± 2%    ~     (p=1.000 n=3+3)
BackendProcessor/minimal-service.ndjson/NoRateLimit-12               18.3µs ± 4%    18.5µs ±10%    ~     (p=1.000 n=3+3)
BackendProcessor/minimal-service.ndjson/WithRateLimit-12             19.2µs ± 4%    19.0µs ± 0%    ~     (p=0.700 n=3+3)
BackendProcessor/minimal.ndjson/NoRateLimit-12                       57.0µs ±14%    47.9µs ±12%    ~     (p=0.400 n=3+3)
BackendProcessor/minimal.ndjson/WithRateLimit-12                     51.1µs ± 9%    50.9µs ± 3%    ~     (p=0.700 n=3+3)
BackendProcessor/optional-timestamps.ndjson/NoRateLimit-12           44.3µs ±10%    45.1µs ± 2%    ~     (p=0.700 n=3+3)
BackendProcessor/optional-timestamps.ndjson/WithRateLimit-12         48.6µs ±15%    46.3µs ±10%    ~     (p=1.000 n=3+3)
BackendProcessor/ratelimit.ndjson/NoRateLimit-12                      119µs ± 2%     119µs ± 2%    ~     (p=1.000 n=3+3)
BackendProcessor/ratelimit.ndjson/WithRateLimit-12                    121µs ± 2%     118µs ± 2%    ~     (p=0.200 n=3+3)
BackendProcessor/spans.ndjson/NoRateLimit-12                          124µs ± 2%     127µs ± 1%    ~     (p=0.100 n=3+3)
BackendProcessor/spans.ndjson/WithRateLimit-12                        124µs ± 1%     128µs ± 2%    ~     (p=0.100 n=3+3)
BackendProcessor/transactions.ndjson/NoRateLimit-12                   134µs ± 2%     133µs ± 2%    ~     (p=0.400 n=3+3)
BackendProcessor/transactions.ndjson/WithRateLimit-12                 136µs ± 1%     136µs ± 1%    ~     (p=1.000 n=3+3)
BackendProcessor/transactions_spans.ndjson/NoRateLimit-12             193µs ± 3%     185µs ± 1%    ~     (p=0.100 n=3+3)
BackendProcessor/transactions_spans.ndjson/WithRateLimit-12           191µs ± 3%     205µs ±17%    ~     (p=1.000 n=3+3)
BackendProcessor/transactions_spans_rum.ndjson/NoRateLimit-12        33.1µs ± 4%    30.1µs ± 3%    ~     (p=0.100 n=3+3)
BackendProcessor/transactions_spans_rum.ndjson/WithRateLimit-12      37.7µs ± 6%    31.7µs ± 1%    ~     (p=0.100 n=3+3)
BackendProcessor/transactions_spans_rum_2.ndjson/NoRateLimit-12      35.6µs ± 7%    31.3µs ± 7%    ~     (p=0.200 n=3+3)
BackendProcessor/transactions_spans_rum_2.ndjson/WithRateLimit-12    35.6µs ± 7%    33.0µs ± 4%    ~     (p=0.200 n=3+3)
RUMV3Processor/rum_errors.ndjson/NoRateLimit-12                      31.5µs ± 8%    27.1µs ± 2%    ~     (p=0.100 n=3+3)
RUMV3Processor/rum_errors.ndjson/WithRateLimit-12                    33.5µs ± 7%    29.5µs ± 4%    ~     (p=0.100 n=3+3)
RUMV3Processor/rum_events.ndjson/NoRateLimit-12                       177µs ± 5%     176µs ± 1%    ~     (p=0.700 n=3+3)
RUMV3Processor/rum_events.ndjson/WithRateLimit-12                     176µs ± 3%     180µs ± 1%    ~     (p=0.400 n=3+3)
[Geo mean]                                                           51.0µs         50.2µs       -1.72%

name                                                               old speed      new speed      delta
BackendProcessor/errors.ndjson/NoRateLimit-12                      47.1MB/s ± 1%  46.3MB/s ± 2%    ~     (p=0.200 n=3+3)
BackendProcessor/errors.ndjson/WithRateLimit-12                    46.4MB/s ± 1%  45.7MB/s ± 3%    ~     (p=1.000 n=3+3)
BackendProcessor/errors_2.ndjson/NoRateLimit-12                    45.8MB/s ± 1%  45.4MB/s ± 1%    ~     (p=0.700 n=3+3)
BackendProcessor/errors_2.ndjson/WithRateLimit-12                  44.9MB/s ± 0%  45.1MB/s ± 3%    ~     (p=0.700 n=3+3)
BackendProcessor/errors_rum.ndjson/NoRateLimit-12                  49.5MB/s ± 9%  58.2MB/s ± 4%    ~     (p=0.100 n=3+3)
BackendProcessor/errors_rum.ndjson/WithRateLimit-12                45.7MB/s ±14%  54.3MB/s ± 3%    ~     (p=0.100 n=3+3)
BackendProcessor/events.ndjson/NoRateLimit-12                      40.1MB/s ± 6%  40.8MB/s ± 0%    ~     (p=1.000 n=3+3)
BackendProcessor/events.ndjson/WithRateLimit-12                    41.3MB/s ± 1%  38.8MB/s ± 4%    ~     (p=0.100 n=3+3)
BackendProcessor/experimental.ndjson/NoRateLimit-12                23.1MB/s ± 7%  26.5MB/s ± 6%    ~     (p=0.100 n=3+3)
BackendProcessor/experimental.ndjson/WithRateLimit-12              21.6MB/s ± 9%  26.1MB/s ± 2%    ~     (p=0.100 n=3+3)
BackendProcessor/heavy.ndjson/NoRateLimit-12                       13.1MB/s ±11%  22.1MB/s ± 2%    ~     (p=0.100 n=3+3)
BackendProcessor/heavy.ndjson/WithRateLimit-12                     13.8MB/s ± 9%  19.8MB/s ± 9%    ~     (p=0.100 n=3+3)
BackendProcessor/invalid-event-type.ndjson/NoRateLimit-12          57.7MB/s ± 1%  51.5MB/s ± 2%    ~     (p=0.100 n=3+3)
BackendProcessor/invalid-event-type.ndjson/WithRateLimit-12        48.1MB/s ± 1%  43.8MB/s ± 1%    ~     (p=0.100 n=3+3)
BackendProcessor/invalid-event.ndjson/NoRateLimit-12               17.3MB/s ± 4%  17.2MB/s ± 5%    ~     (p=1.000 n=3+3)
BackendProcessor/invalid-event.ndjson/WithRateLimit-12             17.0MB/s ± 4%  16.6MB/s ± 1%    ~     (p=0.400 n=3+3)
BackendProcessor/invalid-json-event.ndjson/NoRateLimit-12          32.8MB/s ± 8%  32.1MB/s ± 3%    ~     (p=1.000 n=3+3)
BackendProcessor/invalid-json-event.ndjson/WithRateLimit-12        30.6MB/s ± 3%  29.1MB/s ± 3%    ~     (p=0.100 n=3+3)
BackendProcessor/invalid-json-metadata.ndjson/NoRateLimit-12       23.4MB/s ± 2%  19.7MB/s ±12%    ~     (p=0.100 n=3+3)
BackendProcessor/invalid-json-metadata.ndjson/WithRateLimit-12     23.3MB/s ± 0%  18.1MB/s ±11%    ~     (p=0.100 n=3+3)
BackendProcessor/invalid-metadata-2.ndjson/NoRateLimit-12           115MB/s ± 2%   101MB/s ± 5%    ~     (p=0.100 n=3+3)
BackendProcessor/invalid-metadata-2.ndjson/WithRateLimit-12         115MB/s ± 1%   107MB/s ± 1%    ~     (p=0.100 n=3+3)
BackendProcessor/invalid-metadata.ndjson/NoRateLimit-12             113MB/s ± 1%   104MB/s ± 6%    ~     (p=0.100 n=3+3)
BackendProcessor/invalid-metadata.ndjson/WithRateLimit-12           114MB/s ± 0%    96MB/s ± 6%    ~     (p=0.100 n=3+3)
BackendProcessor/metadata-null-values.ndjson/NoRateLimit-12        41.9MB/s ± 1%  39.6MB/s ± 1%    ~     (p=0.100 n=3+3)
BackendProcessor/metadata-null-values.ndjson/WithRateLimit-12      36.2MB/s ± 2%  35.8MB/s ± 1%    ~     (p=0.400 n=3+3)
BackendProcessor/metadata.ndjson/NoRateLimit-12                     109MB/s ± 1%    99MB/s ± 1%    ~     (p=0.100 n=3+3)
BackendProcessor/metadata.ndjson/WithRateLimit-12                  96.6MB/s ± 0%  87.7MB/s ± 1%    ~     (p=0.100 n=3+3)
BackendProcessor/metricsets.ndjson/NoRateLimit-12                  27.7MB/s ± 1%  27.2MB/s ± 6%    ~     (p=0.700 n=3+3)
BackendProcessor/metricsets.ndjson/WithRateLimit-12                26.9MB/s ± 4%  26.6MB/s ± 2%    ~     (p=1.000 n=3+3)
BackendProcessor/minimal-service.ndjson/NoRateLimit-12             23.2MB/s ± 4%  23.1MB/s ±10%    ~     (p=1.000 n=3+3)
BackendProcessor/minimal-service.ndjson/WithRateLimit-12           22.2MB/s ± 4%  22.4MB/s ± 0%    ~     (p=0.700 n=3+3)
BackendProcessor/minimal.ndjson/NoRateLimit-12                     18.3MB/s ±15%  21.7MB/s ±12%    ~     (p=0.400 n=3+3)
BackendProcessor/minimal.ndjson/WithRateLimit-12                   20.3MB/s ± 9%  20.3MB/s ± 3%    ~     (p=0.700 n=3+3)
BackendProcessor/optional-timestamps.ndjson/NoRateLimit-12         23.3MB/s ±10%  22.8MB/s ± 2%    ~     (p=0.700 n=3+3)
BackendProcessor/optional-timestamps.ndjson/WithRateLimit-12       21.4MB/s ±16%  22.3MB/s ±10%    ~     (p=1.000 n=3+3)
BackendProcessor/ratelimit.ndjson/NoRateLimit-12                   35.3MB/s ± 2%  35.4MB/s ± 2%    ~     (p=1.000 n=3+3)
BackendProcessor/ratelimit.ndjson/WithRateLimit-12                 34.8MB/s ± 2%  35.6MB/s ± 2%    ~     (p=0.200 n=3+3)
BackendProcessor/spans.ndjson/NoRateLimit-12                       34.3MB/s ± 2%  33.4MB/s ± 1%    ~     (p=0.100 n=3+3)
BackendProcessor/spans.ndjson/WithRateLimit-12                     34.4MB/s ± 0%  33.1MB/s ± 2%    ~     (p=0.100 n=3+3)
BackendProcessor/transactions.ndjson/NoRateLimit-12                35.0MB/s ± 2%  35.4MB/s ± 2%    ~     (p=0.400 n=3+3)
BackendProcessor/transactions.ndjson/WithRateLimit-12              34.6MB/s ± 1%  34.5MB/s ± 1%    ~     (p=1.000 n=3+3)
BackendProcessor/transactions_spans.ndjson/NoRateLimit-12          30.1MB/s ± 3%  31.5MB/s ± 1%    ~     (p=0.100 n=3+3)
BackendProcessor/transactions_spans.ndjson/WithRateLimit-12        30.5MB/s ± 3%  28.8MB/s ±16%    ~     (p=1.000 n=3+3)
BackendProcessor/transactions_spans_rum.ndjson/NoRateLimit-12      33.7MB/s ± 4%  37.2MB/s ± 3%    ~     (p=0.100 n=3+3)
BackendProcessor/transactions_spans_rum.ndjson/WithRateLimit-12    29.6MB/s ± 5%  35.2MB/s ± 1%    ~     (p=0.100 n=3+3)
BackendProcessor/transactions_spans_rum_2.ndjson/NoRateLimit-12    30.9MB/s ± 7%  35.1MB/s ± 7%    ~     (p=0.200 n=3+3)
BackendProcessor/transactions_spans_rum_2.ndjson/WithRateLimit-12  30.9MB/s ± 7%  33.2MB/s ± 4%    ~     (p=0.200 n=3+3)
RUMV3Processor/rum_errors.ndjson/NoRateLimit-12                    30.6MB/s ± 9%  35.4MB/s ± 2%    ~     (p=0.100 n=3+3)
RUMV3Processor/rum_errors.ndjson/WithRateLimit-12                  28.7MB/s ± 7%  32.5MB/s ± 4%    ~     (p=0.100 n=3+3)
RUMV3Processor/rum_events.ndjson/NoRateLimit-12                    22.7MB/s ± 5%  22.8MB/s ± 1%    ~     (p=0.700 n=3+3)
RUMV3Processor/rum_events.ndjson/WithRateLimit-12                  22.8MB/s ± 3%  22.4MB/s ± 1%    ~     (p=0.400 n=3+3)
[Geo mean]                                                         34.5MB/s       35.1MB/s       +1.71%

name                                                               old alloc/op   new alloc/op   delta
BackendProcessor/errors.ndjson/NoRateLimit-12                        97.7kB ± 0%    97.8kB ± 0%    ~     (p=0.700 n=3+3)
BackendProcessor/errors.ndjson/WithRateLimit-12                      97.8kB ± 1%    97.9kB ± 0%    ~     (p=0.400 n=3+3)
BackendProcessor/errors_2.ndjson/NoRateLimit-12                      65.5kB ± 0%    65.8kB ± 1%    ~     (p=0.100 n=3+3)
BackendProcessor/errors_2.ndjson/WithRateLimit-12                    65.8kB ± 0%    66.0kB ± 0%    ~     (p=0.100 n=3+3)
BackendProcessor/errors_rum.ndjson/NoRateLimit-12                    18.4kB ± 0%    18.8kB ± 1%    ~     (p=0.100 n=3+3)
BackendProcessor/errors_rum.ndjson/WithRateLimit-12                  18.8kB ± 1%    19.1kB ± 1%    ~     (p=0.100 n=3+3)
BackendProcessor/events.ndjson/NoRateLimit-12                         106kB ± 0%     105kB ± 0%    ~     (p=0.100 n=3+3)
BackendProcessor/events.ndjson/WithRateLimit-12                       105kB ± 0%     106kB ± 0%    ~     (p=0.100 n=3+3)
BackendProcessor/experimental.ndjson/NoRateLimit-12                  19.9kB ± 0%    20.4kB ± 0%    ~     (p=0.100 n=3+3)
BackendProcessor/experimental.ndjson/WithRateLimit-12                20.2kB ± 0%    20.7kB ± 1%    ~     (p=0.100 n=3+3)
BackendProcessor/heavy.ndjson/NoRateLimit-12                         17.1MB ± 0%    17.1MB ± 0%    ~     (p=0.100 n=3+3)
BackendProcessor/heavy.ndjson/WithRateLimit-12                       17.1MB ± 0%    17.2MB ± 0%    ~     (p=0.100 n=3+3)
BackendProcessor/invalid-event-type.ndjson/NoRateLimit-12            2.48kB ± 0%    2.57kB ± 2%    ~     (p=0.100 n=3+3)
BackendProcessor/invalid-event-type.ndjson/WithRateLimit-12          2.85kB ± 1%    2.91kB ± 1%    ~     (p=0.100 n=3+3)
BackendProcessor/invalid-event.ndjson/NoRateLimit-12                 14.5kB ± 0%    14.8kB ± 1%    ~     (p=0.100 n=3+3)
BackendProcessor/invalid-event.ndjson/WithRateLimit-12               14.8kB ± 1%    15.3kB ± 1%    ~     (p=0.100 n=3+3)
BackendProcessor/invalid-json-event.ndjson/NoRateLimit-12            9.48kB ± 1%    9.68kB ± 1%    ~     (p=0.100 n=3+3)
BackendProcessor/invalid-json-event.ndjson/WithRateLimit-12          9.83kB ± 0%   10.01kB ± 1%    ~     (p=0.100 n=3+3)
BackendProcessor/invalid-json-metadata.ndjson/NoRateLimit-12         4.63kB ± 1%    4.66kB ± 2%    ~     (p=1.000 n=3+3)
BackendProcessor/invalid-json-metadata.ndjson/WithRateLimit-12       4.62kB ± 1%    4.62kB ± 0%    ~     (p=1.000 n=3+3)
BackendProcessor/invalid-metadata-2.ndjson/NoRateLimit-12            1.61kB ± 0%    1.62kB ± 1%    ~     (p=0.700 n=3+3)
BackendProcessor/invalid-metadata-2.ndjson/WithRateLimit-12          1.61kB ± 2%    1.64kB ± 0%    ~     (p=0.200 n=3+3)
BackendProcessor/invalid-metadata.ndjson/NoRateLimit-12              1.61kB ± 0%    1.62kB ± 1%    ~     (p=1.000 n=3+3)
BackendProcessor/invalid-metadata.ndjson/WithRateLimit-12            1.62kB ± 0%    1.62kB ± 0%    ~     (p=0.400 n=3+3)
BackendProcessor/metadata-null-values.ndjson/NoRateLimit-12          6.78kB ± 0%    6.92kB ± 0%    ~     (p=0.100 n=3+3)
BackendProcessor/metadata-null-values.ndjson/WithRateLimit-12        7.14kB ± 1%    7.29kB ± 1%    ~     (p=0.100 n=3+3)
BackendProcessor/metadata.ndjson/NoRateLimit-12                      3.15kB ± 1%    3.23kB ± 0%    ~     (p=0.100 n=3+3)
BackendProcessor/metadata.ndjson/WithRateLimit-12                    3.46kB ± 1%    3.59kB ± 0%    ~     (p=0.100 n=3+3)
BackendProcessor/metricsets.ndjson/NoRateLimit-12                    31.7kB ± 1%    32.2kB ± 1%    ~     (p=0.100 n=3+3)
BackendProcessor/metricsets.ndjson/WithRateLimit-12                  32.1kB ± 0%    32.6kB ± 1%    ~     (p=0.100 n=3+3)
BackendProcessor/minimal-service.ndjson/NoRateLimit-12               11.0kB ± 1%    11.1kB ± 1%    ~     (p=0.100 n=3+3)
BackendProcessor/minimal-service.ndjson/WithRateLimit-12             11.3kB ± 1%    11.4kB ± 0%    ~     (p=0.100 n=3+3)
BackendProcessor/minimal.ndjson/NoRateLimit-12                       30.9kB ± 1%    31.3kB ± 1%    ~     (p=0.200 n=3+3)
BackendProcessor/minimal.ndjson/WithRateLimit-12                     31.5kB ± 0%    31.5kB ± 0%    ~     (p=1.000 n=3+3)
BackendProcessor/optional-timestamps.ndjson/NoRateLimit-12           26.2kB ± 1%    26.4kB ± 0%    ~     (p=0.400 n=3+3)
BackendProcessor/optional-timestamps.ndjson/WithRateLimit-12         26.5kB ± 1%    26.6kB ± 0%    ~     (p=1.000 n=3+3)
BackendProcessor/ratelimit.ndjson/NoRateLimit-12                      106kB ± 0%     106kB ± 1%    ~     (p=0.700 n=3+3)
BackendProcessor/ratelimit.ndjson/WithRateLimit-12                    106kB ± 0%     107kB ± 0%    ~     (p=0.200 n=3+3)
BackendProcessor/spans.ndjson/NoRateLimit-12                          109kB ± 0%     110kB ± 0%    ~     (p=0.100 n=3+3)
BackendProcessor/spans.ndjson/WithRateLimit-12                        109kB ± 0%     110kB ± 1%    ~     (p=0.200 n=3+3)
BackendProcessor/transactions.ndjson/NoRateLimit-12                  82.5kB ± 0%    82.6kB ± 0%    ~     (p=0.700 n=3+3)
BackendProcessor/transactions.ndjson/WithRateLimit-12                82.9kB ± 0%    84.5kB ± 0%    ~     (p=0.100 n=3+3)
BackendProcessor/transactions_spans.ndjson/NoRateLimit-12             134kB ± 0%     134kB ± 0%    ~     (p=0.200 n=3+3)
BackendProcessor/transactions_spans.ndjson/WithRateLimit-12           134kB ± 0%     135kB ± 0%    ~     (p=0.100 n=3+3)
BackendProcessor/transactions_spans_rum.ndjson/NoRateLimit-12        18.6kB ± 1%    18.9kB ± 1%    ~     (p=0.100 n=3+3)
BackendProcessor/transactions_spans_rum.ndjson/WithRateLimit-12      18.9kB ± 0%    19.3kB ± 0%    ~     (p=0.100 n=3+3)
BackendProcessor/transactions_spans_rum_2.ndjson/NoRateLimit-12      19.1kB ± 0%    19.5kB ± 1%    ~     (p=0.100 n=3+3)
BackendProcessor/transactions_spans_rum_2.ndjson/WithRateLimit-12    19.5kB ± 1%    19.7kB ± 0%    ~     (p=0.100 n=3+3)
RUMV3Processor/rum_errors.ndjson/NoRateLimit-12                      14.6kB ± 0%    14.8kB ± 0%    ~     (p=0.100 n=3+3)
RUMV3Processor/rum_errors.ndjson/WithRateLimit-12                    15.0kB ± 1%    15.2kB ± 1%    ~     (p=0.200 n=3+3)
RUMV3Processor/rum_events.ndjson/NoRateLimit-12                       132kB ± 0%     134kB ± 0%    ~     (p=0.100 n=3+3)
RUMV3Processor/rum_events.ndjson/WithRateLimit-12                     132kB ± 0%     135kB ± 0%    ~     (p=0.100 n=3+3)
[Geo mean]                                                           26.7kB         27.0kB       +1.25%

name                                                               old allocs/op  new allocs/op  delta
BackendProcessor/errors.ndjson/NoRateLimit-12                         1.15k ± 0%     1.13k ± 0%    ~     (p=0.100 n=3+3)
…